# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MtaaReach CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete/deactivate user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── GEOGRAPHY ─────────────────────────────────────────────────────────────
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts (with search/filter)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: wardId
          in: query
          schema:
            type: integer
        - name: villageId
          in: query
          schema:
            type: integer
        - name: constituencyId
          in: query
          schema:
            type: integer
        - name: countyId
          in: query
          schema:
            type: integer
        - name: groupId
          in: query
          schema:
            type: integer
        - name: gender
          in: query
          schema:
            type: string
        - name: consentSms
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Contacts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactsPage"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Register a new contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "409":
          description: Duplicate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DuplicateContactResponse"
  /contacts/{contactId}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get contact
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update contact
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete contact
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /contacts/{contactId}/opt-out:
    post:
      operationId: optOutContact
      tags: [contacts]
      summary: Opt out contact from SMS
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Opted out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/import:
    post:
      operationId: importContacts
      tags: [contacts]
      summary: Bulk import contacts (CSV)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List contact groups
      responses:
        "200":
          description: Groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group with members
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /groups/{groupId}/members:
    post:
      operationId: addGroupMembers
      tags: [groups]
      summary: Add contacts to group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupMembersInput"
      responses:
        "200":
          description: Members added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: removeGroupMembers
      tags: [groups]
      summary: Remove contacts from group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupMembersInput"
      responses:
        "200":
          description: Members removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── SMS TEMPLATES ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    User:
      type: object
      required: [id, email, firstName, lastName, role, isActive, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [super_admin, tenant_admin, county_coordinator, constituency_coordinator, ward_coordinator, village_coordinator, field_agent]
        isActive:
          type: boolean
        tenantId:
          type: ["integer", "null"]
        assignedCountyId:
          type: ["integer", "null"]
        assignedConstituencyId:
          type: ["integer", "null"]
        assignedWardId:
          type: ["integer", "null"]
        assignedVillageId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    UserUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        assignedCountyId:
          type: ["integer", "null"]
        assignedConstituencyId:
          type: ["integer", "null"]
        assignedWardId:
          type: ["integer", "null"]
        assignedVillageId:
          type: ["integer", "null"]

    # GEOGRAPHY
...
    Contact:
      type: object
      required: [id, firstName, lastName, phone, tenantId, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        gender:
          type: ["string", "null"]
          enum: [male, female, other, null]
        ageGroup:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        countyId:
          type: ["integer", "null"]
        constituencyId:
          type: ["integer", "null"]
        wardId:
          type: ["integer", "null"]
        villageId:
          type: ["integer", "null"]
        pollingStationId:
          type: ["integer", "null"]
        wardName:
          type: ["string", "null"]
        villageName:
          type: ["string", "null"]
        consentSms:
          type: boolean
        consentSource:
          type: ["string", "null"]
        consentDate:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        tenantId:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
    ContactInput:
      type: object
      required: [firstName, lastName, phone]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        ageGroup:
          type: string
        occupation:
          type: string
        countyId:
          type: ["integer", "null"]
        constituencyId:
          type: ["integer", "null"]
        wardId:
          type: ["integer", "null"]
        villageId:
          type: ["integer", "null"]
        pollingStationId:
          type: ["integer", "null"]
        consentSms:
          type: boolean
        consentSource:
          type: string
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
    ContactUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        gender:
          type: ["string", "null"]
        ageGroup:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        wardId:
          type: ["integer", "null"]
        villageId:
          type: ["integer", "null"]
        pollingStationId:
          type: ["integer", "null"]
        consentSms:
          type: boolean
        tags:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
    ContactsPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    DuplicateContactResponse:
      type: object
      required: [error, existing]
      properties:
        error:
          type: string
        existing:
          $ref: "#/components/schemas/Contact"
    ContactImportInput:
      type: object
      required: [contacts]
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/ContactInput"
    ImportResult:
      type: object
      required: [imported, skipped, errors]
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
...
    Group:
      type: object
      required: [id, name, tenantId, memberCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        tenantId:
          type: integer
        memberCount:
          type: integer
        createdAt:
          type: string
    GroupInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    GroupUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
    GroupMembersInput:
      type: object
      required: [contactIds]
      properties:
        contactIds:
          type: array
          items:
            type: integer

    # TEMPLATES
...